remove all shadow property because it can be implemented by shader.
